+++ demo/ButtonGroupDemo.jsx
@@ -15,6 +15,7 @@ class Demo extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
+      maxLength: 1,
     };
   }