+++ src/Separated.jsx
@@ -5,6 +5,7 @@ import Button from 'uxcore-button';
 import Dropdown from 'uxcore-dropdown';
 import Menu from 'uxcore-menu';
 import classnames from 'classnames';
+import { polyfill } from 'react-lifecycles-compat';
 import i18n from './i18n';
 
 /**
@@ -20,6 +21,7 @@ class Separated extends React.Component {
     onClick: PropTypes.func,
     size: PropTypes.string,
   };
+
   static defaultProps = {
     maxLength: 3,
     locale: 'zh-cn',
@@ -27,6 +29,16 @@ class Separated extends React.Component {
     actionType: 'button',
     size: 'medium',
   };
+
+  static getDerivedStateFromProps = (props) => 
+    // if (React.Children.count(props.children) < parseInt(props.maxLength, 10)) {
+    //   return {
+    //     dropdownVisible: false,
+    //   };
+    // }
+     null
+  
+
   constructor(props) {
     super(props);
     this.state = {
@@ -35,14 +47,6 @@ class Separated extends React.Component {
     this.handleDropdownVisibleChange = this.handleDropdownVisibleChange.bind(this);
     this.handleMoreClick = this.handleMoreClick.bind(this);
     this.refCallback = [];
-  }
-
-  componentWillReceiveProps(nextProps) {
-    if (React.Children.count(nextProps.children) < parseInt(nextProps.maxLength, 10)) {
-      this.setState({
-        dropdownVisible: false,
-      });
-    }
   }
 
   componentDidUpdate() {
@@ -55,7 +59,9 @@ class Separated extends React.Component {
       /* eslint-disable react/no-find-dom-node */
       const triggerDOMnode = ReactDOM.findDOMNode(this.triggerInstance);
       /* eslint-enable react/no-find-dom-node */
-      dropdownDOMNode.style.minWidth = `${(triggerDOMnode || this.triggerInstance).offsetWidth}px`;
+      if (dropdownDOMNode) {
+        dropdownDOMNode.style.minWidth = `${(triggerDOMnode || this.triggerInstance).offsetWidth}px`;
+      }
     }
   }
 
@@ -321,5 +327,7 @@ class Separated extends React.Component {
 }
 
 Separated.displayName = 'Separated';
+
+polyfill(Separated);
 
 export default Separated;