+++ HISTORY.md
@@ -1,5 +1,10 @@
 # history
 
+## 0.2.0
+
+* `CHANGED` update `uxcore-dropdown` to `0.5.0`
+* `CHANGED` adapt React@16
+
 ## 0.1.6
 
 * `CHANGED` update `uxcore-menu` to `^3.0.0`